Typo fix in schema migration operator

+++ src/gcp_airflow_foundations/operators/gcp/schema_migration/schema_migration_operator.py
@@ -198,7 +198,7 @@ class MigrateSchema(BaseOperator):
 
     def bigQuery_mapping(self, data_type):
         mapping = {
-            "FLOAT":"FLOAT64"
+            "FLOAT":"FLOAT"
         }
 
         if data_type in mapping:
